many MANY many typos fixed added proper animation supported - can register global animation for each entity type - can reuse this data across animation components reviewed and cleaned up asset loading (AssetManager)
renamed raylib methods to avoid name conflict
made texture atlas user defined
- now supports unlimited amounts

+++ src/core/StaticCollision.cpp
@@ -82,10 +82,9 @@ namespace magique
         {
             if (markedClassMap.contains(tileInfo.getClass()))
             {
-                data.markedTilesMap.insert(tileInfo.tileID+1);
+                data.markedTilesMap.insert(static_cast<uint16_t>(tileInfo.tileID + 1));
             }
         }
-
     }
 
     void LoadTileMap(const MapID map, const TileMap& tileMap, const std::initializer_list<int>& layers)
@@ -132,6 +131,5 @@ namespace magique
     }
 
     void RemoveColliderGroup(int group) {}
-
 
 } // namespace magique
